# style: Get rid of allocations in class selector matching