--- a/style/BUILD
+++ b/style/BUILD
@@ -5,7 +5,10 @@ cc_library(
     name = "style",
     srcs = glob(
         include = ["*.cpp"],
-        exclude = ["*_test.cpp"],
+        exclude = [
+            "*_test.cpp",
+            "*_bench.cpp",
+        ],
     ),
     hdrs = glob(["*.h"]),
     copts = HASTUR_COPTS,
@@ -33,3 +36,16 @@ cc_library(
         "//gfx",
     ],
 ) for src in glob(["*_test.cpp"])]
+
+[cc_test(
+    name = src.removesuffix(".cpp"),
+    size = "small",
+    srcs = [src],
+    copts = HASTUR_COPTS,
+    deps = [
+        ":style",
+        "//dom",
+        "//etest",
+        "@nanobench",
+    ],
+) for src in glob(["*_bench.cpp"])]