Before:
```
|               ns/op |                op/s |    err% |     total | is_match: class
|--------------------:|--------------------:|--------:|----------:|:----------------
|             1392.13 |           718322.70 |    1.2% |      0.01 | `match, few classes`
|             2272.80 |           439986.22 |    0.6% |      0.01 | `no match, few classes`
|             6605.88 |           151380.23 |    1.8% |      0.01 | `match, many classes`
|             5102.84 |           195969.17 |    0.2% |      0.01 | `no match, many classes`
```

After removing the bonus allocations:
```
|               ns/op |                op/s |    err% |     total | is_match: class
|--------------------:|--------------------:|--------:|----------:|:----------------
|              341.30 |          2929955.00 |    0.5% |      0.01 | `match, few classes`
|              536.39 |          1864327.70 |    0.7% |      0.01 | `no match, few classes`
|             1595.52 |           626753.07 |    1.3% |      0.01 | `match, many classes`
|             1160.63 |           861598.44 |    0.2% |      0.01 | `no match, many classes`
```

After also not looking up the class attribute once per class searched for:
```
|               ns/op |                op/s |    err% |     total | is_match: class
|--------------------:|--------------------:|--------:|----------:|:----------------
|              297.82 |          3357777.59 |    0.6% |      0.01 | `match, few classes`
|              452.88 |          2208102.43 |    1.0% |      0.01 | `no match, few classes`
|             1444.01 |           692515.43 |    0.4% |      0.01 | `match, many classes`
|             1085.37 |           921346.23 |    0.6% |      0.01 | `no match, many classes`
```

There's more low-hanging fruit, but dealing with that will be nicer now that there's easy access to benchmarking.